Change variable names for alphabetical order
When running `terraform`, variables are prompted for on the command line. The AWS secret key and access key need to be entered, but for the rest, for now, the default can be accepted. This change puts the secret key and access key up first so you can just enter them and then press return through the rest of them, for ease of use.