Add link to Terraform CLI documentation
I was going to add links directly to `plan` and `apply`, which is how we are running it at the moment, but actually we are all so new to Terraform that we might find a better way and I don't want this documentation to go out of date. The link to the CLI doc is useful in itself as it's not obvious where to start.