Copy app-servers for GCE
Contrary to the name this contains instances and load balancers for:

- API
- Router
- Gandalf

We should split these out for both at a later date.

Not using auto scaling groups because they aren't supported by Terraform for
GCE. We're currently unsure about whether we're going to continue using auto
scaling groups for AWS because we haven't integrated them with Ansible's
config management.

Any tags necessary for firewall rules will be fixed in a following commit.