Add shell script to wrap GCE DNS
Terraform doesn't have a resource type for Google's Cloud DNS service. So
for the moment we need to call out to the `gcloud` command line utility.

However it doesn't itself provide a simple way to ensure that a record is
set to the value that we want. So we need to wrap it with a shell script
that deletes the existing record of the same name, regardless of it's
attributes, if it already exists.