Check error message response from gcloud command line
If there is already a transaction pending, then the current
one will fail... but it doesn't tell us this until we try
to execute the transaction.

We parse the error message back from the gcloud command to
look for code: 412 messages (Prerequisites not met) and
should we find that message, we have to abort the current
transaction and create a new one from the start.