Remove health check from Router LB in GCE
The routers are difficult to health check because they don't respond to
requests themselves. Everything gets proxied to something else below. For a
brand new install it doesn't have any routes configured and for an existing
setup we don't have an app that we can rely on being there.

GCE only supports HTTP and HTTPS health checks on "forwarding rules". We
might have more options for "load balancers" but they aren't currently
supported by Terraform. So remove the health checks for now and always send
traffic to members.