aws: Centralise and restrict office IPs
Use the variable which is shared by GCE for the list of office IPs. This
also restricts it to the two egress IPs that we own, rather than the entire
range which is used by other customers of our provider.

Some rules have been de-duped because we can pass a list of addresses.