Avoid using cloud-clonfig for Autoscaling instances
To be consistent with GCE, which is not using cloud-config.
The commands ran in user-data are being executed by the Ansible provisioner anyway.
See also https://github.com/alphagov/tsuru-ansible/pull/27

Next phase will be to remove cloud-config/user-data from single instances (Gandalf, Docker) as well, but that requires resource replacement, so we need to be plan accordingly.