aws: Replace ASGs with instances+count
Auto-scaling Groups have been problematic for us for two reasons:

- we can't easily use the launch configuration to run Ansible on instance
  creation because we're using delegation and it would require open SSH
  access between machines
- they're not supported by Terraform on GCE so we have disparity between the
  two sets of configs

So replace them with normal instances and `count` to create multiples of
them. Now that Terraform knows about the instances themselves we will be
able to use its state file to drive Ansible's inventory. I've added
additional outputs of these instances like we have for GCE so that we can
input them into the inventory for the timebeing.

We may later periodically run Terraform to ensure that we have the desired
state and we could dynamically adjust the counts.

In it's current form this puts all of the instances into the same
availability zone (default is the first). I'll address this in a following
commit. It's not trivial to do because subnets are also tied to AZs.

I've kept the launch configs and moved them to a separate file, indicating
that they should later be removed, because Terraform doesn't know about the
relationships when destroying unmanaged resources from the statefile and was
failing with:

    aws_launch_configuration.tsuru-sslproxy: Destroying...
    aws_launch_configuration.api: Destroying...
    aws_launch_configuration.router: Destroying...
    Error applying plan:

    3 error(s) occurred:

    * Cannot delete launch configuration terraform-Sz66CQVUUQF3rN9tyDP5xdLPkX0= because it is attached to AutoScalingGroup dcarley-tsuru-api-asg
    * Cannot delete launch configuration terraform-lj1WhKZSBN8NMqI9bgdYZ7OsYzc= because it is attached to AutoScalingGroup dcarley-tsuru-sslproxy-asg
    * Cannot delete launch configuration terraform-lpuvad7RYsnJrKSNRq2S7ckT09Q= because it is attached to AutoScalingGroup dcarley-tsuru-router-asg