Add unique identifier for dynamic environments in GCE
Previously, we were not able to create different environments from the same template.
This was due to resource naming conflicts.
We are now making the resource names unique per environement. E.g: `tsuru-router-dev`, `tsuru-router-prod`
The environment name is passed to Terraform as a parameter when creating the infrastructure, e.g:
`terraform apply -var 'env=prod'`