aws: Replace separate subnet resources with count
Create two of each kind of subnet, one for each availability zone, by using
`count` instead of separate resources. This is a pre-req for assigning
instances to separate AZs based on their count because we'll need to lookup
(probably using `element()`) the corresponding subnet and zone.

NB: This will cause most instances to be recreated because a subnet can't be
changed on an existing instance.

I did try using `"${var.region}a"` in the `zones` variable but Terraform
doesn't allow you to interpolate variables within variables at assignment.