aws: Split instances across availability zones
These instances were previously in auto-scaling groups and would have been
distributed across the two availability zones that we have been using (a and
b). This commits restores that by looking up the appropriate zone and subnet
based on the `count`. The `element()` function will wrap so a third instance
will be assigned to the first zone and so on.