Add detail about shared `account.json`
This is context that was missing in the set-up, for people on our team new to this repo.

This commit also splits out the two extra requirements into a numbered list, for clarity. When we incorporate the [Terrafom change](https://www.pivotaltracker.com/n/projects/1275640/stories/93451408) that means we do not have to shell out to `gcloud`, the second one can be removed.