Add root ssh directory to gitignore
GCE requires its own ssh directory as per AWS but is also [looking in the directory above](https://github.com/alphagov/tsuru-terraform/blob/master/gce/variables.tf#L27) for the SSH key. It should be consiistent with AWS but I will do that in a separate PR. For now, just adding the ssh directory to gitignore so that it doesn't accidentally get committed.