Get rid of need for ssh directory in root folder
The previous set up required the ssh key (and directory) both in the root folder and in the gce folder. This is probably a hangover from before the infrastructure was divided into aws and gce. This commit changes where the variables file is looking for the SSH key so the ssh folder in the root directory is not required and can be deleted by anyone who has it.

This commit also removes the entry from `gitignore` which I added in a [previous PR](https://github.com/alphagov/tsuru-terraform/pull/42) to avoid accidentally committing it.