aws: Move all "private" records to "public" zone
The one existing private zone (referred to by zone ID) needed to be
associated with each new independent VPC. Terraform isn't able to do this
because it doesn't manage that existing private zone and doing it by hand
would be too cumbersome. This was documented in the story but not the
README.

This meant that machines were unable to resolve the addresses of internal
ELBs and failed to configure Nginx, in the example of the following Ansible
inventory item:

    [hipache]
    xx.xx.xx.xx internal_ip=dcarley-hipache-int.tsuru.paas.alphagov.co.uk

We could fix this by letting Terraform create and manage a separate private
zone for each independent environment. Since these don't have to be
delegated like public zones do.

However we're not using the split-horizon feature of private zones (our
records are namespaced to be "internal" and don't overlap with the public
records) and I'm not concerned about us exposing private IP addresses.

The current private zone will have to be deleted manually.