aws/gce: Move trailing dot to dns_zone_name
This is needed by `google_dns_record_set` resources otherwise GCE's API will
complain that the record is invalid. Moving it from each of the resources to
the variable means less repetition and chance for mistakes. I'm changing AWS
to match, but it has no functional effect because Route53 doesn't care if
it's there or not.