gce: Combine sslproxy and router roles
Combine the Nginx which does TLS termination and reverse-proxying for the
routers onto the same machines as the routers. The separate `sslproxy`
machines and their associated resources are no longer needed. This means
that:

- the address of the upstream doesn't change like DNS records for ELBs do
- we're not subject to extra network latency/partitions for reverse proxying
- we have a simpler architecture

I've had to FIXME some resources which can't be deleted in a single
Terraform run. We'll need to remove these later after this has been deployed
to any existing environments like CI.

There will be an accompanying PR to alphagov/tsuru-ansible which will need
to be merged at the same time.