aws: Remove router-int load balancer
This was used by Ansible on AWS to configure the Nginx sslproxy machines to
reverse-proxy to the routers. The hostname was referenced in the variable
`hipache_host_internal_lb`. We no longer need this since moving Nginx to the
same machines as the router.

Although we might want to revisit this again later when we test one
application talking to another application through the router - which we
wouldn't want to pay external traffic for.