Remove one hop when resolving app name
Because of legacy configuration (2 LBs, one for SSL and one for HTTP) there is leftover configuration which redirects *.<env>-hipache.<DOMAIN> to either SSL or HTTP LBs - whichever you configure. We are now only using SSL. Therefore it's not necessary for this DNS record to point to <env>-hipache.<DOMAIN> which then points to the actual SSL Load Balancer, but we can point to the LB directly. This simplifies config and speeds up DNS resolution (one step less to resolve final LB IP). Switch to SSL only configuration was made in https://github.com/alphagov/tsuru-ansible/pull/69