Create CNAME for <env>-postgresqlapi.<DOMAIN>
To instanciate postgresql service instance, tsuru api needs to talk to the postgresqlapi app (which is postgresql broker). This happens over https. We only have certificate for *.<DOMAIN>, which doesn't match postgresqlapi.<env>-hipache.<DOMAIN>. To work around this, we are creating a CNAME record <env>-postgresqlapi.<DOMAIN> which points to the platform LB directly (same target as for *.<env>-hipache.<DOMAIN>). We also map this route to the postgresqlapi app inside Tsuru, so that hipache can route requests correctly. After this change, we register postgresqlapi broker on the <env>-postgresqlapi.<DOMAIN> URL. Tsuru API now talks to this URL which matches the *.<DOMAIN> certificate, hence all works.