Make ELB's forward tcp/443 versus terminate SSL
We don't currently have SSL on the API in GCE environments because the
load balancers which do SSL termination aren't currently supported by
Terraform.

We do have ELBs terminating SSL in AWS, but the certificate upload
process is outside of Terraform.

In order to harmonise how we deal with SSL termination across both
platforms (`gce` and `aws`) and ensure we can fully automate the
generation of environments we will forward the TCP traffic to a
[Nginx](http://wiki.nginx.org/Main) `SSL` terminator on the `tsuru api server`.

This commit consists of:

* [Changing the AWS load balancer to forward to TCP:443 instead of terminating SSL](https://github.com/alphagov/tsuru-ansible/pull/74)
* Changing the healthcheck to check on `HTTPS:443/info` of the `tsuru api server`