Add the HTTPS load balancing for api in gce
We need to setup a new `google_compute_forwarding_rule` in order
to provide HTTPS load balancing for the api running on gce.

In this commit we:

* Created a new check, as we need to specify a different port: 443
* Setup a new pool, as we need to use a different check
* Create the forwarding rule, referencing the new pool. It reuses
  the `google_compute_address` assigned to the 8080 rule.

We added a `https` suffix to all the resources, as done in the
`router` resources, but we keep the names for the old
resources to avoid conflicts applying terraform.