Remove the still unsupported gce HTTPS healthcheck
Currently HTTPS health checks in GCE are part of the GCE beta API and are not
supported by terraform.

There's a [feature request](https://github.com/hashicorp/terraform/issues/2190)
where these possible implementations are discussed:

 1. Alternate resource type: google_compute_https_health_check
 2. Add a new attribute `protocol`

In this commit remove the HTTPS check and use only HTTP on port 8080, for
both the `HTTP/8080` and `HTTPS/443` forwarding rules. The drawbacks are:

 * We don't check the nginx SSL terminator service in the health check.
 * The insecure API must be open, at least for the LB layer.