aws: Create IAM role for registry on S3
Create an IAM role and instance profile, associated with the registry
instance, which will allow it to access S3 without using a predefined
access/secret key.

The policy matches what we had previously setup by hand for the single
username/keypair. It will allow access to the buckets of all independent
environments. It's not possible to restrict this further without giving
everyone access to IAM.

NB: The `iam_instance_profile` attribute is `ForceNew` so the registry
instance will be recreated. This is OK though, because the data is in S3 and
the index will be rebuilt from S3 when the new machine comes up.