force_destroy parameter for GCE and AWS buckets
Avoid the following error:
Error deleting S3 Bucket: BucketNotEmpty: The bucket you tried to delete is not empty
	status code: 409, request id: []

Allows deleting the content of S3 and GCE buckets on a terraform destroy action.
This is really useful when using test environments. It is potentially dangerous as the bucket resource may be implicitly destroyed and recreated during a simple apply.
It can be specified via the force_destroy variable, which is false by default.