Update README to explain apply limiting target
To change the force_deploy setting in the buckets you need to run apply again
limiting the target to the buckets. It is better to set the variable
during the initial provisioning.

This commit explains how set this variable in the initial provision and
how to change the setting if it was not set previously.