Make AWS NAT server run on ubuntu
Make the AWS NAT server run on ubuntu, just like the GCE NAT and all other hosts. This way we can have unified ssh.config (used in tsuru-ansible repository), which can be used to ssh to all hosts - including the NAT server itself. This was previously not possible, as the user name on this server differed (ec2-user vs ubuntu).