Change the script to setup NAT routing servers
There were a pair of issues in the configuration of NAT routing. First,
the inline commands in terraform will not stop the execution if they
fail, you have to use an external script for that.

Second, while installing a package at boot time with the provisioner,
cloud-init might be running and it changes the upstream sources.list.
Because that we add logic to wait for cloud-init to finish.