aws/gce: Remove all outputs
We were originally using these to manually create Ansible's inventory file.
It was possible that we were going to re-use these to programatically
create/source an inventory file, but dynamic inventories using the AWS and
GCE APIs turned out to be better. So we no longer have a use for these and
we haven't been very consistent at updating them.