Revert "gitignore: add ETCD_CLUSTER_ID files"
This reverts commit 42c16dccf97af032cb69e81f364ebfa460fc3bfb.

This wasn't having the desired effect. git won't ignore changes to files
that are also tracked. So I'm removing these to avoid confusion.

Because Terraform requires these files to be present on the first run (so we
can't not commit them) we'll just have to be careful about not committing
any changes back, for now:

    Errors:

      * 1 error(s) occurred:

    * 1 error(s) occurred:

    * file: open ETCD_CLUSTER_ID: no such file or directory in:

    ${file("ETCD_CLUSTER_ID")}