Makefile prep target to create ETCD_CLUSTER_ID
These files need to be present (but empty) the first time Terraform runs
otherwise it will generate the error (which I've documented in the README).

We can't commit the empty files to git because, even with `.gitignore`, they
will still be tracked and show as being modified. Which means that there's a
risk we commit the IDs back and it makes changing branches difficult.

So I'm adding a make target which creates the empty files and documenting
that it has to be run before Terraform. By using `touch` it won't affect the
contents of the file if they already exist - although it wouldn't be a
problem if it did, because the content of the file is cached in Terraform's
statefile.