Create a user called 'core' in the NAT server
We need a user called 'core' in the NAT server, in order to be able
to use fleet with `$FLEETCTL_SSH_USERNAME` and `$FLEETCTL_TUNNEL`.

When you run `fleet journal` with `$FLEETCTL_TUNNEL`, it will connect
first the nat server but then the target servers using the user
`$FLEETCTL_SSH_USERNAME`. Because that, you need the same user in all
the boxes.