Added gandalf and graffana to nat-to-internal.
Adding new target tags allows SSH via NAT.

We tried by adding the tag private to both the `gandalf` and `grafana`
servers, but this breaks one of the connections.

For instance, in gandalf, you could SSH from the NAT server, but SSH to
the public interface of the gandalf server would fail. `tcpdump` show
that the traffic was reaching and going out of the server, but the
response never arrived to the client. This behaviour suggest some kind
of bug in Google Compute Engine firewall, that does not allow returning
traffic if a host matches two rules with different tags.

We should investigate this issue in the future. Other approach might be
setup TCP loadbalancers for the public services in this servers (SSH and
grafana web interface), instead of exposing the server directly.