Add systemd unit to run telegraf in CoreOS
Add a systemd unit in the metadata passed to CoreOS to start a docker
container which runs telegraf. This container mounts the system
directories and sends the metrics to influxdb.

The IP of influxdb is populated by terraform.

The user and password and DB name are hardcoded for the time being.