Set grafana tags for docker coreos machines
Include the tags `job` as the name of the AWS or GCE instance, and type
as a job type label (`docker-coreos`).

Now we need to create a specific user-data template for each machine, so
we add a `count` to the template resource. We gather the instance name
from the template variables.