Updated readme with extra step for destruction.
Due to a [bug in Terraform](https://github.com/hashicorp/terraform/issues/2957)
removal of disk attachements is attempted whilst the volume is still in use.
To work around this bug, it is necessary to first `taint` the instance before
destroying it. When the instance is tainted, terraform doesn't wait for the
volume to be unused before destruction.